// src/app/interfaces/CryptoJS.ts
export interface CryptoJSInterface{
    out: string;
    wordArray : string;

    userSesion: string;
    keyValue: string;
    key: string;
    iv: string;
    //textencryipted: string;
    //encrypt2Value: string;

    data:string;

    rawData:string;
    crypttext:string;

}

export interface Vencriptdecript{
    encrypt: string;
    decrypt: string;

}